The PDF 1st edition
Any thoughts on what should be added would be greatly appreciated.  The target audience for this set of PDF sheets is newer Linux users, people interested in using the CLI, and students.  The grungy graphics and text are only to promote interest and to make it feel like a DIY underground punk rock project but I want the content to be correct and useful.